Renamed classes, moved the to better package
Removed the package "switcher" because it was confusing
