+++ subprojects/deprecatedPluginsTest/src/test/resources/mockito-extensions/org.mockito.plugins.InstantiatorProvider
@@ -1,1 +1,1 @@
-org.mockitousage.plugins.switcher.MyInstantiatorProvider
+org.mockitousage.plugins.MyDeprecatedInstantiatorProvider
